Fix Flight , Airport , Plane and Image url

+++ src/main/java/com/binar/bejticketing/service/Flight/PlaneDetailsServiceImpl.java
@@ -1,13 +1,16 @@
 package com.binar.bejticketing.service.Flight;
 
 
+import com.binar.bejticketing.entity.AgeCategory;
 import com.binar.bejticketing.entity.PlaneDetails;
+import com.binar.bejticketing.exception.DataNotFoundException;
 import com.binar.bejticketing.repository.PlaneDetailsRepository;
 import com.binar.bejticketing.service.PlaneDetailsService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class PlaneDetailsServiceImpl implements PlaneDetailsService {
@@ -22,12 +25,26 @@ public class PlaneDetailsServiceImpl implements PlaneDetailsService {
 
     @Override
     public PlaneDetails updatePlaneDetails(PlaneDetails planeDetails) {
-        return planeDetailsRepository.save(planeDetails);
+        Optional<PlaneDetails> byId = planeDetailsRepository.findById(planeDetails.getIdPlaneClass());
+
+        if (byId.isPresent()){
+            return planeDetailsRepository.save(planeDetails);
+        }
+        throw new DataNotFoundException(planeDetails.getIdPlaneClass());
     }
 
     @Override
-    public void deletePlaneDetails(Long id) {
-        planeDetailsRepository.deleteById(id);
+    public String deletePlaneDetails(Long id) {
+        Optional<PlaneDetails> byId = planeDetailsRepository.findById(id);
+
+        if (byId.isPresent()){
+            planeDetailsRepository.deleteById(id);
+            return "Delete Success";
+        }
+
+        throw  new DataNotFoundException(id);
+
+
     }
