+++ src/main/java/com/binar/bejticketing/service/Flight/PlaneServiceImpl.java
@@ -1,6 +1,8 @@
 package com.binar.bejticketing.service.Flight;
 
 import com.binar.bejticketing.entity.Plane;
+import com.binar.bejticketing.entity.PlaneDetails;
+import com.binar.bejticketing.exception.DataNotFoundException;
 import com.binar.bejticketing.repository.PlaneRepository;
 import com.binar.bejticketing.service.PlaneService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,12 +24,24 @@ public class PlaneServiceImpl implements PlaneService {
 
     @Override
     public Plane updatePlane(Plane plane) {
-        return planeRepository.save(plane);
+        Optional<Plane> byId = planeRepository.findById(plane.getIdPlane());
+
+        if (byId.isPresent()){
+            return planeRepository.save(plane);
+        }
+        throw new DataNotFoundException(plane.getIdPlane());
     }
 
     @Override
-    public void deletePlane(Long id) {
-        planeRepository.deleteById(id);
+    public DataNotFoundException deletePlane(Long id) {
+        Optional<Plane> byId = planeRepository.findById(id);
+
+        if (byId.isPresent()){
+            planeRepository.deleteById(id);
+            return null;
+        }
+
+        return new DataNotFoundException(id);
     }
 
     @Override
@@ -37,6 +51,7 @@ public class PlaneServiceImpl implements PlaneService {
 
     @Override
     public Optional<Plane> getPlaneById(Long id) {
+
         return planeRepository.findById(id);
     }
 }
